refactor(frontend): Avoid loop import for type GetFeeData
